Move MeetingTarget to correct folder
