+++ src/main/java/seedu/address/logic/commands/meetingcommands/DeleteMeetingCommand.java
@@ -6,6 +6,7 @@ import java.util.Objects;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.parser.MeetingTarget;
 import seedu.address.model.Model;
 import seedu.address.model.meeting.Meeting;
