+++ src/main/java/seedu/address/logic/commands/meetingcommands/DeleteMeetingsTagCommand.java
@@ -10,6 +10,7 @@ import java.util.Set;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.parser.MeetingTarget;
 import seedu.address.model.Model;
 import seedu.address.model.meeting.Duration;
 import seedu.address.model.meeting.Link;
