+++ src/main/java/seedu/address/logic/parser/MeetingTarget.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.meetingcommands;
+package seedu.address.logic.parser;
 
 import static java.util.Objects.requireNonNull;
 
@@ -9,7 +9,6 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.meeting.Meeting;
 import seedu.address.model.meeting.Title;
