+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -11,7 +11,6 @@ import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
-import seedu.address.logic.commands.meetingcommands.MeetingTarget;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.meeting.Duration;
 import seedu.address.model.meeting.Link;
