+++ src/main/java/seedu/address/logic/parser/meetingcommands/AddMeetingCommandParser.java
@@ -14,9 +14,9 @@ import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.add.AddTagOnlyCommand;
 import seedu.address.logic.commands.meetingcommands.AddMeetingCommand;
 import seedu.address.logic.commands.meetingcommands.AddTagToMeetingCommand;
-import seedu.address.logic.commands.meetingcommands.MeetingTarget;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.parser.MeetingTarget;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.Prefix;
