+++ src/main/java/seedu/address/logic/parser/meetingcommands/DeleteMeetingCommandParser.java
@@ -7,9 +7,9 @@ import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.delete.DeleteTagOnlyCommand;
 import seedu.address.logic.commands.meetingcommands.DeleteMeetingCommand;
 import seedu.address.logic.commands.meetingcommands.DeleteMeetingsTagCommand;
-import seedu.address.logic.commands.meetingcommands.MeetingTarget;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.parser.MeetingTarget;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
