+++ src/test/java/seedu/address/storage/JsonSerializableMeetingsBookTest.java
@@ -1,4 +1,4 @@
-package seedu.meeting.storage;
+package seedu.address.storage;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.address.testutil.Assert.assertThrows;
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.JsonUtil;
 import seedu.address.model.MeetingsBook;
-import seedu.address.storage.JsonSerializableMeetingsBook;
 import seedu.address.testutil.TypicalMeetings;
 
 public class JsonSerializableMeetingsBookTest {
